refactoring test functions
